+++ silo-oracles/contracts/dia/DIAOracle.sol
@@ -53,12 +53,7 @@ contract DIAOracle is ISiloOracle, IDIAOracle, Initializable {
         if (_baseToken != data.baseToken) revert AssetNotSupported();
         if (_baseAmount > type(uint128).max) revert BaseAmountOverflow();
 
-        (
-            uint128 assetPrice,
-            bool priceUpToDate
-        ) = getPriceForKey(data.diaOracle, primaryKey[cacheOracleConfig], data.heartbeat);
-
-        if (!priceUpToDate) revert OldPrice();
+        uint128 assetPrice = getPriceForKey(data.diaOracle, primaryKey[cacheOracleConfig]);
 
         if (!data.convertToQuote) {
             quoteAmount = OracleNormalization.normalizePrice(
@@ -69,11 +64,7 @@ contract DIAOracle is ISiloOracle, IDIAOracle, Initializable {
             return quoteAmount;
         }
 
-        (
-            uint128 secondaryPrice, bool secondaryPriceValid
-        ) = getPriceForKey(data.diaOracle, secondaryKey[cacheOracleConfig], data.heartbeat);
-
-        if (!secondaryPriceValid) revert OldSecondaryPrice();
+        uint128 secondaryPrice = getPriceForKey(data.diaOracle, secondaryKey[cacheOracleConfig]);
 
         quoteAmount = OracleNormalization.normalizePrices(
             _baseAmount,
@@ -99,25 +90,17 @@ contract DIAOracle is ISiloOracle, IDIAOracle, Initializable {
 
     /// @param _diaOracle IDIAOracleV2 oracle where price is stored
     /// @param _key string under this key asset price will be available in DIA oracle
-    /// @param _heartbeat period after which price became invalid
     /// @return assetPriceInUsd uint128 asset price
-    /// @return priceUpToDate bool TRUE if price is up to date (acceptable), FALSE otherwise
-    function getPriceForKey(IDIAOracleV2 _diaOracle, string memory _key, uint256 _heartbeat)
+    function getPriceForKey(IDIAOracleV2 _diaOracle, string memory _key)
         public
         view
         virtual
-        returns (uint128 assetPriceInUsd, bool priceUpToDate)
+        returns (uint128 assetPriceInUsd)
     {
         uint128 priceTimestamp;
         (assetPriceInUsd, priceTimestamp) = _diaOracle.getValue(_key);
         if (priceTimestamp == 0) revert InvalidKey();
 
-        // price must be updated at least once every 24h, otherwise something is wrong
-        uint256 oldestAcceptedPriceTimestamp;
-        // block.timestamp is more than HEARTBEAT, so we can not underflow
-        unchecked { oldestAcceptedPriceTimestamp = block.timestamp - _heartbeat; }
-
         // we not checking assetPriceInUsd != 0, because this is checked on setup, so it will be always some value here
-        priceUpToDate = priceTimestamp > oldestAcceptedPriceTimestamp;
     }
 }