+++ silo-oracles/test/foundry/dia/DIAOracle.t.sol
@@ -27,19 +27,6 @@ contract DIAOracleTest is DIAConfigDefault {
         DIA_ORACLE.initialize(cfg, _defaultDIAConfig().primaryKey, _defaultDIAConfig().secondaryKey);
     }
 
-    function test_DIAOracle_initialize_OldPrice() public {
-        DIAOracle newOracle = DIAOracle(Clones.clone(address(new DIAOracle())));
-        IDIAOracle.DIADeploymentConfig memory cfg = _defaultDIAConfig(10 ** (18 + 8 - 18), 0);
-
-        cfg.heartbeat = 1856;
-        DIAOracleConfig newConfig = new DIAOracleConfig(cfg);
-
-        newOracle.initialize(newConfig, cfg.primaryKey, cfg.secondaryKey);
-
-        vm.expectRevert(IDIAOracle.OldPrice.selector);
-        newOracle.quote(1e18, address(tokens["RDPX"]));
-    }
-
     function test_DIAOracle_initialize_OldSecondaryPrice() public {
         DIAOracle newOracle = DIAOracle(Clones.clone(address(new DIAOracle())));
         IDIAOracle.DIADeploymentConfig memory cfg = _defaultDIAConfig(10 ** (18 + 8 - 18), 0);