+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Removed
+- silo-oracles: remove heartbeat check from DIA oracle
 
 ### [3.7.0] - 2025-06-10
 ### Added